[DAEF-68] refactor components related to wallet transactions
