+++ app/components/wallet/transactions/WalletNoTransactions.js
