+++ app/components/wallet/transactions/WalletNoTransactions.scss
