+++ app/components/wallet/transactions/WalletTransactionsList.scss
