+++ app/containers/wallet/WalletTransactionsPage.js
@@ -2,9 +2,9 @@
 import React, { Component, PropTypes } from 'react';
 import { observer, inject, PropTypes as MobxPropTypes } from 'mobx-react';
 import { defineMessages, intlShape } from 'react-intl';
-import WalletTransactionsList from '../../components/wallet/home/WalletTransactionsList';
+import WalletTransactionsList from '../../components/wallet/transactions/WalletTransactionsList';
 // import WalletTransactionsSearch from '../../components/wallet/home/WalletTransactionsSearch';
-import WalletNoTransactions from '../../components/wallet/home/WalletNoTransactions';
+import WalletNoTransactions from '../../components/wallet/transactions/WalletNoTransactions';
 import CachedRequest from '../../stores/lib/CachedRequest';
 
 const messages = defineMessages({
